requirements.txt
EasyProcess==0.3
entrypoint2==0.2.1
ffmpeg-python==0.2.0
freetype-py==2.1.0.post1
future==0.18.2
gif2numpy==1.3
kaitaistruct==0.8
mss==5.1.0
numpy==1.18.4
numpy2gif==1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
Pillow==7.1.2
pyee==7.0.1
pyscreenshot==1.0
